requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.2.16
langchain-google-genai==1.0.10
pydantic==2.7.4
python-dotenv==1.0.0
python-multipart==0.0.6
aiohttp==3.9.0
